requirements.txt
colorama==0.4.6
gradio==4.44.0
playwright==1.47.0
python-dotenv==1.0.1
